requirements.txt
llama-index
python-dotenv
html2text
llama-index-readers-web

-e .